more deletions + some renaming
