+++ .env.example
@@ -1,36 +1,9 @@
-# bitcoin node running in your private network 192.168.1.0/24
-BITCOIN_RPC_URL=http://192.168.1.100
-
-# bitcoin node running undockered on the same PC
-# needs to add rpcallowip=172.16.0.0/12 to your bitcoin.conf
-#BITCOIN_RPC_URL=http://host.docker.internal
-
-BITCOIN_RPC_USER=
-BITCOIN_RPC_PASSWORD=
-BITCOIN_RPC_PORT=8332
-BITCOIN_RPC_TIMEOUT=10000
-
-# You can use this instead of BITCOIN_RPC_USER and BITCOIN_RPC_PASSWORD
-BITCOIN_RPC_COOKIEFILE=
-
-# Enable in bitcoin.conf with
-# zmqpubrawblock=tcp://*:3000
-# BITCOIN_ZMQ_HOST="tcp://192.168.1.100:3000"
-
 API_PORT=3334
 STRATUM_PORT=3333
 
-#optional telegram bot
-#TELEGRAM_BOT_TOKEN=
-
-#optional discord bot
-#DISCORD_BOT_CLIENTID=
-#DISCORD_BOT_GUILD_ID=
-#DISCORD_BOT_CHANNEL_ID=
-
-#optional
-DEV_FEE_ADDRESS=
-# mainnet | testnet
-NETWORK=mainnet
+# "https://api.mainnet-beta.solana.com" | "https://api.devnet.solana.com" | <custom>
+SOLANA_URL="http://localhost:8899" # test validator url
+# "processed" | "confirmed" | "finalized" | "recent" | "single" | "singleGossip" | "root" | "max"
+SOLANA_COMMITMENT="confirmed" 
 
 API_SECURE=false
